main/chat/paginators.py
from rest_framework.pagination import PageNumberPagination


class ChatPagination(PageNumberPagination):
    page_size = 30
    page_size_query_param = 'page_size'
    max_page_size = 60


class RequestPagination(PageNumberPagination):
    page_size = 40
    page_size_query_param = 'page_size'
    max_page_size = 60
