requirements.txt
Django==3.2
djangorestframework==3.14.0
Markdown==3.4.4
django-filter==23.2
psycopg2-binary==2.8.6
Pillow==10.0.0
drf-yasg==1.21.7
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
xmltodict==0.13.0
dicttoxml==1.7.16
django-filter==23.2
django-multiselectfield==0.1.12
django-jazzmin==2.6.0
django-admin-sortable2==1.0.4
celery==5.3.6
redis==5.0.1
django-solo==2.1.0
firebase-admin==6.3.0
django-environ==0.11.2
gunicorn==21.2.0
django-celery-beat==2.5.0
django-cleanup==8.1.0
sentry-sdk==1.44.0
fcm-django==2.1.0
django-jet-4==0.0.1